requirements.txt
requests
beautifulsoup4
lxml
google-search-results
thefuzz[speedup]
python-dateutil
tenacity
jinja2
